requirements.txt
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.7.0
requests==2.32.3
boto3==1.34.141
google-search-results==2.4.2
PyJWT[crypto]==2.9.0
httpx==0.27.2
cachetools==5.5.0
celery==5.4.0
redis==5.0.7
